requirements.txt
mysql-connector-python
networkx
python-louvain
numpy
scikit-learn
matplotlib